+++ CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.14](https://github.com/Salable/salable-web-components-stenciljs/compare/v0.1.13...v0.1.14) (2024-01-24)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [0.1.13](https://github.com/Salable/salable-web-components-stenciljs/compare/v0.1.12...v0.1.13) (2024-01-24)
+
+
+### Bug Fixes
+
+* add rough outline of docs ([#66](https://github.com/Salable/salable-web-components-stenciljs/issues/66)) ([681c206](https://github.com/Salable/salable-web-components-stenciljs/commit/681c20627f4a2298b0fadd1f575d0a10c47ece13))
+
+
+
+
+
+## [0.1.12](https://github.com/Salable/salable-web-components-stenciljs/compare/v0.1.11...v0.1.12) (2024-01-24)
+
+
+### Bug Fixes
+
+* clean up and align styles + set styles to scoped for checkout ([#65](https://github.com/Salable/salable-web-components-stenciljs/issues/65)) ([2b55712](https://github.com/Salable/salable-web-components-stenciljs/commit/2b5571244ab4a262c31f98e1b668899f35c100f7))
+
+
+
+
+
 ## [0.1.11](https://github.com/Salable/salable-web-components-stenciljs/compare/v0.1.10...v0.1.11) (2024-01-22)